+++ dispatchable_generator/lib/src/dispatchable_generator.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
 
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:build/src/builder/build_step.dart';
 import 'package:dispatchable/dispatchable.dart';
 import 'package:meta/meta.dart';
@@ -16,37 +18,100 @@ class DispatchableGenerator extends Generator {
   FutureOr<String> generate(LibraryReader library, BuildStep buildStep) {
     return library
         .annotatedWith(_DispatchableAnnotationTypeChecker)
-        .map((e) => generateForElement(e.element, library))
+        .map((e) => generateForElement(e.element, e.annotation, library))
         .join();
   }
 
   @visibleForTesting
-  String generateForElement(Element element, LibraryReader library) {
-    ClassElement classElement = _validateElement(element);
-    List<ClassElement> subClasses =
-        DispatchableGenerator._findAllSubClassesInFile(library, classElement);
+  String generateForElement(
+      Element element, ConstantReader annotation, LibraryReader library) {
+    var classes = annotation.read('classes').listValue;
+    ClassElement classElement = _validateElement(element, false);
+    var subTypes = classes.isNotEmpty
+        ? _findSubTypesForAllClasses(classes, library)
+        : [_findSubTypesForSingleClass(library, classElement)];
+    subTypes.forEach((e) => _validateSubType(e, classes.isNotEmpty));
+
     DispatchableClassGenerator dispatchableCodeBuilder =
-        DispatchableClassGenerator.validateAndCreate(classElement, subClasses);
+        DispatchableClassGenerator.create(
+            classElement,
+            subTypes,
+            annotation.read('methodSeparator').stringValue,
+            annotation.read('methodPrefix').stringValue);
     return [
       dispatchableCodeBuilder.generateDispatcherClass(),
       dispatchableCodeBuilder.generateDefaultDispatcherClass(),
     ].join("\n");
   }
 
-  static ClassElement _validateElement(Element element) {
-    if (element is! ClassElement) {
+  static List<TypeWithSubTypes> _findSubTypesForAllClasses(
+      List<DartObject> classes, LibraryReader library) {
+    List<TypeWithSubTypes> subTypes = [];
+    for (var x in classes) {
+      var typeValue = x.toTypeValue();
+      ClassElement classElement = _validateAnnotationClass(typeValue);
+      var subClasses = _findAllSubClassesInFileFromType(library, typeValue);
+      subTypes.add(TypeWithSubTypes(classElement, subClasses));
+    }
+    return subTypes;
+  }
+
+  static ClassElement _validateAnnotationClass(DartType? type) {
+    if (type == null) {
       throw InvalidGenerationSourceError(
-          "@dispatchable can only be used to annotate a class.",
-          todo: "Remove @dispatchable annotation from the offending element.",
-          element: element);
+        "@Dispatchable's classes parameter only supports classes.",
+        todo: "Remove the null value from @Dispatchable's classes parameter.",
+      );
+    }
+    if (type.element == null) {
+      throw InvalidGenerationSourceError(
+          "@Dispatchable's classes parameter only supports classes.",
+          todo: "Remove ${type} from @Dispatchable's classes parameter.");
+    }
+    return _validateElement(type.element!, false);
+  }
+
+  static ClassElement _validateElement(Element element, bool multiClassMode) {
+    if (element is! ClassElement || element.isEnum) {
+      var error = multiClassMode
+          ? "@Dispatchable's classes parameter only supports classes."
+          : "@Dispatchable only supports classes.";
+      var todo = multiClassMode
+          ? "Remove ${element.name} from @Dispatchable's classes parameter."
+          : "Remove @Dispatchable from ${element.name}.";
+      throw InvalidGenerationSourceError(error, todo: todo, element: element);
     }
     return element;
   }
 
-  static List<ClassElement> _findAllSubClassesInFile(
+  static void _validateSubType(
+      TypeWithSubTypes typeWithSubTypes, bool multiClassMode) {
+    if (typeWithSubTypes.type.isAbstract && typeWithSubTypes.subTypes.isEmpty) {
+      var todo = multiClassMode
+          ? 'Remove ${typeWithSubTypes.type.name} from the classes list'
+          : 'Remove @Dispatchable from ${typeWithSubTypes.type.name}';
+      throw InvalidGenerationSourceError(
+          '@Dispatchable does not support abstract classes with no sub '
+          'classes.',
+          todo: todo + ' or define sub classes for it.',
+          element: typeWithSubTypes.type);
+    }
+  }
+
+  static TypeWithSubTypes _findSubTypesForSingleClass(
       LibraryReader libraryReader, ClassElement element) {
+    return TypeWithSubTypes(
+        element,
+        libraryReader.classes
+            .where((s) => s.supertype?.element == element)
+            .toList());
+  }
+
+  static List<ClassElement> _findAllSubClassesInFileFromType(
+      LibraryReader libraryReader, DartType? element) {
     return libraryReader.classes
-        .where((s) => s.supertype?.element == element)
+        .where((s) =>
+            s.supertype != null && s.supertype.hashCode == element.hashCode)
         .toList();
   }
 }