+++ dispatchable/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1-dev.3
+- @Dispatchable now accepts a list of classes for which it will generate all possible
+  combinations of sub-types for each class in turn, super useful for use with Bloc.
+- @Dispatchable can now be configured to use provided method prefixes and separators.
+
 ## 0.0.1-dev.2
 - Updated example generated code to match the new version of dispatchable_generator