+++ dispatchable_generator/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1-dev.3 
+- Support building dispatchers for a list of classes.
+- Prefix generated classes with _$
+- Rename static dispatch function to always be acceptFunc.
+
 ## 0.0.1-dev.2
 - No longer generates null checks as dart's null soundness now provides that for us. 
 - Generate code using single quotes instead of double to avoid dart analysis warnings.